Update logic to handle item picture
The case is that the vendor picture that is passed to the method is the current picture, hence this leads to the issue where the initialItem is equal to the uploadedFile. Additional conditions are added to make sure that the operation only takes place when the pictures are different.

+++ src/main/java/backend/file_io/PictureIO.java
@@ -106,13 +106,14 @@ public class PictureIO {
         // Retrieve the item picture
         File initialItem = Utility.retrieveFileWithoutExtension(directory, fileName);
 
-        // Remove the picture and return true
-        if (initialItem != null) {
+        // Remove the picture and return true (if the initial picture and uploaded file is different)
+        // Picture will be same if no files are uploaded
+        if (initialItem != null && !initialItem.equals(uploadedFile)) {
             if (!initialItem.delete()) return false;
         }
 
-        // Save the file if the picture is not null
-        if (uploadedFile != null) {
+        // Save the file if the picture is not null and the picture is not the same picture
+        if (uploadedFile != null && !uploadedFile.equals(initialItem)) {
 
             // Generate file name
             String[] initialFileName = uploadedFile.getName().split("\\.");
